requirements.txt
requests~=2.32.3
PyQt5~=5.15.11
python-dotenv~=1.0.1
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0